requirements.txt
pyserial
numpy
progress
svgoutline
lxml
opencv-python